refactor: split export.ts into format-specific modules (#1396)
