+++ src/lib/utils/export/index.ts
@@ -1,0 +1,14 @@
+/**
+ * Export utilities for generating images from rack layouts.
+ *
+ * This barrel re-exports the public API previously provided by export.ts so
+ * existing `$lib/utils/export` import sites resolve unchanged.
+ */
+
+export { generateExportSVG, generateSingleRackSVG } from "./svg";
+export { exportAsSVG, exportAsPDF, prepareSvgForPdf } from "./vector";
+export { exportAsPNG, exportAsJPEG } from "./raster";
+export { exportAsZip, exportAsMultiPagePDF } from "./multi";
+export type { ExportProgressCallback } from "./multi";
+export { exportToCSV } from "./data";
+export { downloadBlob, generateExportFilename } from "./utils";
