+++ src/lib/utils/export/svg.ts
@@ -7,30 +7,24 @@ import type {
   RackGroup,
   DeviceType,
   ExportOptions,
-  ExportFormat,
   DeviceCategory,
-  ExportBackground,
 } from "$lib/types";
 import type { ImageStoreMap } from "$lib/types/images";
-import { getBlockedSlots } from "./blocked-slots";
+import { getBlockedSlots } from "../blocked-slots";
 import {
   fitTextToWidth,
   DEVICE_LABEL_MAX_FONT,
   DEVICE_LABEL_MIN_FONT,
   DEVICE_LABEL_ICON_SPACE_LEFT,
   DEVICE_LABEL_ICON_SPACE_RIGHT,
-} from "./text-sizing";
+} from "../text-sizing";
 import {
   U_HEIGHT_PX,
   BASE_RACK_WIDTH,
   RAIL_WIDTH,
   RACK_PADDING_HIDDEN,
 } from "$lib/constants/layout";
-import { toHumanUnits, formatPosition } from "$lib/utils/position";
-
-// Note: jsPDF and svg2pdf.js are imported dynamically in PDF export functions
-// to avoid loading the large jsPDF bundle (~200KB) on app startup.
-// See issue #68 for details.
+import { toHumanUnits } from "$lib/utils/position";
 
 // Aliases for export context (export uses hidden padding since view labels show rack name)
 const U_HEIGHT = U_HEIGHT_PX;
@@ -1478,227 +1472,6 @@ export function generateExportSVG(
 }
 
 /**
- * Export SVG element as string with XML declaration
- */
-export function exportAsSVG(svg: SVGElement): string {
-  const serializer = new XMLSerializer();
-  const svgString = serializer.serializeToString(svg);
-  return `<?xml version="1.0" encoding="UTF-8"?>\n${svgString}`;
-}
-
-/**
- * Export SVG as PNG blob
- */
-export async function exportAsPNG(
-  svg: SVGElement,
-  scale: number = 2,
-): Promise<Blob> {
-  return exportAsRaster(svg, "image/png", scale);
-}
-
-/**
- * Export SVG as JPEG blob
- */
-export async function exportAsJPEG(
-  svg: SVGElement,
-  scale: number = 2,
-  quality: number = 0.92,
-): Promise<Blob> {
-  return exportAsRaster(svg, "image/jpeg", scale, quality);
-}
-
-/**
- * Normalise an export SVG so svg2pdf.js renders it faithfully.
- *
- * svg2pdf does not implement every SVG/CSS feature the browser does, so the
- * shared export SVG needs two adjustments before vector PDF conversion (these
- * are applied to the re-parsed copy only — the PNG path keeps the original):
- *
- *  - `dominant-baseline` is ignored by svg2pdf; it reads `alignment-baseline`
- *    instead. Mirror the value across so vertically-centred text (device names,
- *    U numbers) stays centred rather than shifting up to the baseline (#1738).
- *  - Numeric `font-weight` values other than 400/700 yield an invalid jsPDF
- *    style (e.g. "normal600") and fall back to the serif font (times). Collapse
- *    weights to the supported 400/700, treating >= 500 as bold (#1739).
- */
-export function prepareSvgForPdf(svgElement: Element): void {
-  const textLike = [
-    ...Array.from(svgElement.getElementsByTagName("text")),
-    ...Array.from(svgElement.getElementsByTagName("tspan")),
-  ];
-
-  for (const el of textLike) {
-    const dominantBaseline = el.getAttribute("dominant-baseline");
-    if (dominantBaseline && !el.getAttribute("alignment-baseline")) {
-      el.setAttribute("alignment-baseline", dominantBaseline);
-    }
-
-    const fontWeight = el.getAttribute("font-weight");
-    if (fontWeight) {
-      const isBold =
-        fontWeight === "bold" ||
-        fontWeight === "bolder" ||
-        Number(fontWeight) >= 500;
-      el.setAttribute("font-weight", isBold ? "700" : "400");
-    }
-  }
-}
-
-/**
- * Export SVG string as PDF blob (US Letter size, centered)
- */
-export async function exportAsPDF(
-  svgString: string,
-  // Background is embedded in the SVG element; vector PDF conversion preserves it directly
-  _background: ExportBackground,
-): Promise<Blob> {
-  const { jsPDF } = await import("jspdf");
-  await import("svg2pdf.js");
-
-  const parser = new DOMParser();
-  const svgDoc = parser.parseFromString(svgString, "image/svg+xml");
-  const svgElement = svgDoc.documentElement;
-
-  // svg2pdf lacks support for some SVG features the browser renders; normalise
-  // this re-parsed copy so the vector PDF matches the on-screen/PNG output.
-  prepareSvgForPdf(svgElement);
-
-  const imgWidth = parseInt(svgElement.getAttribute("width") || "0", 10);
-  const imgHeight = parseInt(svgElement.getAttribute("height") || "0", 10);
-
-  if (imgWidth === 0 || imgHeight === 0) {
-    throw new Error("Invalid SVG dimensions");
-  }
-
-  // US Letter dimensions in points (72 dpi)
-  const letterWidth = 612; // 8.5 inches
-  const letterHeight = 792; // 11 inches
-
-  // Create PDF (landscape if image is wider than tall)
-  const isLandscape = imgWidth > imgHeight;
-
-  const pdf = new jsPDF({
-    orientation: isLandscape ? "landscape" : "portrait",
-    unit: "pt",
-    format: "letter",
-  });
-
-  // Calculate scaling to fit on page with margins
-  const margin = 36; // 0.5 inch margins
-  const pageWidth = isLandscape ? letterHeight : letterWidth;
-  const pageHeight = isLandscape ? letterWidth : letterHeight;
-  const availableWidth = pageWidth - margin * 2;
-  const availableHeight = pageHeight - margin * 2;
-
-  const scale = Math.min(
-    availableWidth / imgWidth,
-    availableHeight / imgHeight,
-  );
-
-  const scaledWidth = imgWidth * scale;
-  const scaledHeight = imgHeight * scale;
-
-  // Center on page
-  const x = (pageWidth - scaledWidth) / 2;
-  const y = (pageHeight - scaledHeight) / 2;
-
-  // Convert SVG to vector PDF commands
-  await pdf.svg(svgElement, { x, y, width: scaledWidth, height: scaledHeight });
-
-  return pdf.output("blob");
-}
-
-/**
- * Internal function to render SVG to canvas and export as raster image
- */
-async function exportAsRaster(
-  svg: SVGElement,
-  mimeType: string,
-  scale: number,
-  quality?: number,
-): Promise<Blob> {
-  const width = parseInt(svg.getAttribute("width") || "0", 10);
-  const height = parseInt(svg.getAttribute("height") || "0", 10);
-
-  // Create canvas
-  const canvas = document.createElement("canvas");
-  canvas.width = width * scale;
-  canvas.height = height * scale;
-
-  const ctx = canvas.getContext("2d");
-  if (!ctx) {
-    throw new Error("Failed to get canvas context");
-  }
-
-  // Scale the canvas
-  ctx.scale(scale, scale);
-
-  // For JPEG, fill with white background first (no transparency)
-  if (mimeType === "image/jpeg") {
-    ctx.fillStyle = "#ffffff";
-    ctx.fillRect(0, 0, width, height);
-  }
-
-  // Convert SVG to data URL
-  const svgString = new XMLSerializer().serializeToString(svg);
-  const svgBlob = new Blob([svgString], {
-    type: "image/svg+xml;charset=utf-8",
-  });
-  const url = URL.createObjectURL(svgBlob);
-
-  try {
-    // Load image
-    const img = await loadImage(url);
-
-    // Draw to canvas
-    ctx.drawImage(img, 0, 0);
-
-    // Convert to blob
-    return new Promise<Blob>((resolve, reject) => {
-      canvas.toBlob(
-        (blob) => {
-          if (blob) {
-            resolve(blob);
-          } else {
-            reject(new Error("Failed to create blob"));
-          }
-        },
-        mimeType,
-        quality,
-      );
-    });
-  } finally {
-    URL.revokeObjectURL(url);
-  }
-}
-
-/**
- * Load an image from URL
- */
-function loadImage(url: string): Promise<HTMLImageElement> {
-  return new Promise((resolve, reject) => {
-    const img = new Image();
-    img.onload = () => resolve(img);
-    img.onerror = () => reject(new Error("Failed to load image"));
-    img.src = url;
-  });
-}
-
-/**
- * Trigger download of a blob
- */
-export function downloadBlob(blob: Blob, filename: string): void {
-  const url = URL.createObjectURL(blob);
-  const link = document.createElement("a");
-  link.href = url;
-  link.download = filename;
-  document.body.appendChild(link);
-  link.click();
-  document.body.removeChild(link);
-  URL.revokeObjectURL(url);
-}
-
-/**
  * Generate an SVG element for a single rack
  * This is a simpler version of generateExportSVG for multi-rack exports
  *
@@ -1714,262 +1487,4 @@ export function generateSingleRackSVG(
   images?: ImageStoreMap,
 ): SVGElement {
   return generateExportSVG([rack], deviceLibrary, options, images);
-}
-
-/**
- * Progress callback for multi-rack exports
- */
-export type ExportProgressCallback = (progress: {
-  current: number;
-  total: number;
-  rackName: string;
-}) => void;
-
-/**
- * Export multiple racks as a ZIP file containing one image per rack
- *
- * @param racks - Racks to export
- * @param deviceLibrary - Device library for device definitions
- * @param options - Export options (format must be png, jpeg, or svg)
- * @param images - Optional map of device images
- * @param onProgress - Optional progress callback
- * @returns Promise resolving to ZIP blob
- */
-export async function exportAsZip(
-  racks: Rack[],
-  deviceLibrary: DeviceType[],
-  options: ExportOptions,
-  images?: ImageStoreMap,
-  onProgress?: ExportProgressCallback,
-): Promise<Blob> {
-  // Dynamically import JSZip wrapper
-  const { createZip, generateRackFilename } = await import("./zip");
-
-  const format = options.format === "jpeg" ? "jpg" : options.format;
-  const files: { name: string; blob: Blob }[] = [];
-
-  for (const [i, rack] of racks.entries()) {
-    onProgress?.({ current: i + 1, total: racks.length, rackName: rack.name });
-
-    // Generate SVG for this rack
-    const svg = generateSingleRackSVG(rack, deviceLibrary, options, images);
-
-    // Convert to the requested format
-    let blob: Blob;
-    if (options.format === "svg") {
-      const svgString = exportAsSVG(svg);
-      blob = new Blob([svgString], { type: "image/svg+xml" });
-    } else if (options.format === "png") {
-      blob = await exportAsPNG(svg);
-    } else if (options.format === "jpeg") {
-      blob = await exportAsJPEG(svg);
-    } else {
-      throw new Error(`Unsupported format for ZIP export: ${options.format}`);
-    }
-
-    // Generate filename for this rack
-    const view = options.exportView || "both";
-    const filename = generateRackFilename(rack.name, view, format);
-
-    // Ensure unique filenames by appending index if duplicate
-    const existingNames = files.map((f) => f.name);
-    let finalName = filename;
-    let counter = 1;
-    while (existingNames.includes(finalName)) {
-      const ext = filename.split(".").pop();
-      const base = filename.replace(`.${ext}`, "");
-      finalName = `${base}-${counter}.${ext}`;
-      counter++;
-    }
-
-    files.push({ name: finalName, blob });
-  }
-
-  return createZip(files);
-}
-
-/**
- * Export SVG string as multi-page PDF (one rack per page)
- *
- * @param racks - Racks to export (one per page)
- * @param deviceLibrary - Device library for device definitions
- * @param options - Export options
- * @param images - Optional map of device images
- * @param onProgress - Optional progress callback
- * @returns Promise resolving to PDF blob
- */
-export async function exportAsMultiPagePDF(
-  racks: Rack[],
-  deviceLibrary: DeviceType[],
-  options: ExportOptions,
-  images?: ImageStoreMap,
-  onProgress?: ExportProgressCallback,
-): Promise<Blob> {
-  // Dynamically import jsPDF to avoid loading it on app startup
-  const { jsPDF } = await import("jspdf");
-  await import("svg2pdf.js");
-
-  // US Letter dimensions in points (72 dpi)
-  const letterWidth = 612; // 8.5 inches
-  const letterHeight = 792; // 11 inches
-  const margin = 36; // 0.5 inch margins
-  const headerHeight = 30; // Space for rack name header
-
-  let pdf: InstanceType<typeof jsPDF> | null = null;
-
-  for (const [i, rack] of racks.entries()) {
-    onProgress?.({ current: i + 1, total: racks.length, rackName: rack.name });
-
-    // Generate SVG for this rack
-    const svg = generateSingleRackSVG(rack, deviceLibrary, options, images);
-
-    // Parse SVG to get dimensions
-    const imgWidth = parseInt(svg.getAttribute("width") || "0", 10);
-    const imgHeight = parseInt(svg.getAttribute("height") || "0", 10);
-
-    if (imgWidth === 0 || imgHeight === 0) {
-      console.warn(
-        `Skipping rack "${rack.name}" due to invalid SVG dimensions (${imgWidth}x${imgHeight})`,
-      );
-      continue;
-    }
-
-    // Determine page orientation based on rack aspect ratio
-    const isLandscape = imgWidth > imgHeight;
-    const pageWidth = isLandscape ? letterHeight : letterWidth;
-    const pageHeight = isLandscape ? letterWidth : letterHeight;
-    const availableWidth = pageWidth - margin * 2;
-    const availableHeight = pageHeight - margin * 2 - headerHeight;
-
-    // Calculate scaling to fit on page
-    const scale = Math.min(
-      availableWidth / imgWidth,
-      availableHeight / imgHeight,
-    );
-    const scaledWidth = imgWidth * scale;
-    const scaledHeight = imgHeight * scale;
-
-    // Center on page
-    const x = (pageWidth - scaledWidth) / 2;
-    const y = margin + headerHeight + (availableHeight - scaledHeight) / 2;
-
-    // Initialize PDF or add new page
-    if (pdf === null) {
-      pdf = new jsPDF({
-        orientation: isLandscape ? "landscape" : "portrait",
-        unit: "pt",
-        format: "letter",
-      });
-    } else {
-      pdf.addPage("letter", isLandscape ? "landscape" : "portrait");
-    }
-
-    // Add rack name as header
-    pdf.setFontSize(14);
-    pdf.setFont("helvetica", "bold");
-    pdf.text(rack.name, pageWidth / 2, margin + 20, { align: "center" });
-
-    // Normalise SVG for svg2pdf before conversion (baseline + font-weight).
-    // This SVG is generated per page and not shared with the PNG path.
-    prepareSvgForPdf(svg);
-
-    // Convert SVG to vector PDF
-    await pdf.svg(svg, { x, y, width: scaledWidth, height: scaledHeight });
-  }
-
-  if (!pdf) {
-    throw new Error("No valid racks to export");
-  }
-
-  return pdf.output("blob");
-}
-
-/**
- * Escape a CSV field value
- * - Wraps in quotes if contains comma, quote, or newline
- * - Doubles any existing quotes
- */
-function escapeCSVField(value: string): string {
-  if (value.includes(",") || value.includes('"') || value.includes("\n")) {
-    return `"${value.replace(/"/g, '""')}"`;
-  }
-  return value;
-}
-
-/**
- * Export rack contents as CSV
- * Columns: Position, Name, Model, Manufacturer, U_Height, Category, Face
- * Sorted by position descending (top of rack first)
- *
- * @param rack - The rack to export
- * @param deviceTypes - Device type library for resolving device details
- */
-export function exportToCSV(rack: Rack, deviceTypes: DeviceType[]): string {
-  const header = "Position,Name,Model,Manufacturer,U_Height,Category,Face";
-
-  // Create a map for quick device type lookup
-  const deviceTypeMap = new Map(deviceTypes.map((dt) => [dt.slug, dt]));
-
-  // Sort devices by position descending (top of rack first)
-  const sortedDevices = [...rack.devices].sort(
-    (a, b) => b.position - a.position,
-  );
-
-  // Build rows
-  const rows: string[] = [];
-  for (const device of sortedDevices) {
-    const deviceType = deviceTypeMap.get(device.device_type);
-    if (!deviceType) continue; // Skip unknown device types
-
-    const position = formatPosition(device.position);
-    const name = escapeCSVField(device.name || "");
-    const model = escapeCSVField(deviceType.model || deviceType.slug);
-    const manufacturer = escapeCSVField(deviceType.manufacturer || "");
-    const uHeight = String(deviceType.u_height);
-    const category = deviceType.category;
-    const face = device.face;
-
-    rows.push(
-      `${position},${name},${model},${manufacturer},${uHeight},${category},${face}`,
-    );
-  }
-
-  return [header, ...rows].join("\n");
-}
-
-/**
- * Generate a sanitized filename for export
- * Pattern: {layout-name}-{view}-{YYYY-MM-DD}.{ext}
- * For CSV (view=null): {layout-name}-{YYYY-MM-DD}.{ext}
- *
- * @param layoutName - The layout name to include in filename
- * @param view - The export view ('front', 'rear', 'both') or null for data exports like CSV
- * @param format - The export format extension
- */
-export function generateExportFilename(
-  layoutName: string,
-  view: "front" | "rear" | "both" | null,
-  format: ExportFormat,
-): string {
-  // Slugify the layout name: lowercase, hyphens, no special chars
-  const slugified = layoutName
-    .toLowerCase()
-    .replace(/[^a-z0-9]+/g, "-")
-    .replace(/^-+|-+$/g, "");
-
-  const baseName = slugified || "Rackula-export";
-
-  // Format date as YYYY-MM-DD
-  const now = new Date();
-  const year = now.getFullYear();
-  const month = String(now.getMonth() + 1).padStart(2, "0");
-  const day = String(now.getDate()).padStart(2, "0");
-  const dateStr = `${year}-${month}-${day}`;
-
-  // Build filename: include view for image exports, omit for CSV
-  if (view) {
-    return `${baseName}-${view}-${dateStr}.${format}`;
-  }
-
-  return `${baseName}-${dateStr}.${format}`;
 }
